feat: modularize smart contract and fix test compilation and event errors
